Twitter crawler with a little bit of spice
In this task i crawl through #newyork and #dog hashtags on twitter, collect the hashtags that appear in those two main hashtags.
From there i do a breadth first search and use the collected hashtags to collect hashtags from these too.

With all of these i create a graph and add nodes and edges between these hashtags. The result of added nodes and edges, with a little bit of detail fixing is seen in the newyorkdog.pdf file.